Address requirements composition review feedback
Keep requirements_composition/mod.rs focused on module wiring and exports, moving the input-layer preparation into layer.rs and the stack/fold logic into stack.rs. This mirrors the ConfigLayerEntry/ConfigLayerStack split and makes the public input unit RequirementsLayerEntry.

Move special domain merge policies into their own modules: rules.rs for additive prefix rules, hooks.rs for append-only hook events plus managed-dir conflict handling, and permissions.rs for additive permissions.filesystem.deny_read. Convert module-level policy comments to //! docs.

Tighten the AbsolutePathBufGuard lifetime so it only wraps layer parsing, use expression style for non-early returns, and switch stack-level tests to the #[path] pattern with stack_tests.rs. Expected requirement outputs are now mostly expressed as TOML parsed into ConfigRequirementsToml and compared as whole objects with pretty_assertions, while provenance-specific assertions stay explicit.
